- Improved SoundTouch::flush() so that it produces exactly accurate number of output samples. - Changed 'float' variables into 'double' for more precise calculation of input-vs-output samples.

+++ source/SoundTouch/InterpolateCubic.h
@@ -56,7 +56,7 @@ protected:
                         const SAMPLETYPE *src, 
                         int &srcSamples);
 
-    float fract;
+    double fract;
 
 public:
     InterpolateCubic();
